+++ sixtyfps_compiler/passes/ensure_window.rs
@@ -18,7 +18,8 @@ use std::collections::HashSet;
 use std::rc::Rc;
 
 pub fn ensure_window(component: &Rc<Component>, type_register: &TypeRegister) {
-    if component.root_element.borrow().base_type.to_string() == "Window" {
+    if matches!(component.root_element.borrow().base_type.to_string().as_str(), "Window" | "Dialog")
+    {
         return; // already a window, nothing to do
     }